Finished implementing comp_tiempo.c
There were many changes, some of these are
- Finished debugging the comp tiempo calculations
- Changed all calculations from float to double types
- Cleaned the process of initialization

Not much more missing as the TODO states

+++ include/rendimientos.h
@@ -13,19 +13,19 @@
 // TODO: cambiar list[] por nombres mas descriptivos
 
 // Function prototypes
-float TREL_EXPORT volumen_combustible(engine_t *engine);
-float TREL_EXPORT masa_combustible(engine_t* engine);
-float TREL_EXPORT Tiempo_quemado_combustion(engine_t* engine);
+double TREL_EXPORT volumen_combustible(engine_t *engine);
+double TREL_EXPORT masa_combustible(engine_t* engine);
+double TREL_EXPORT Tiempo_quemado_combustion(engine_t* engine);
 double TREL_EXPORT Ab_combustion(engine_t* engine);
-float TREL_EXPORT br_combustion(engine_t* engine);
+double TREL_EXPORT br_combustion(engine_t* engine);
 double TREL_EXPORT mg_combustion(engine_t* engine);
 double TREL_EXPORT emp_teo(engine_t* engine);
 double TREL_EXPORT emp_prom(trel_rocket_t* rocket);
 double TREL_EXPORT emp_max(trel_rocket_t* rocket);
-float TREL_EXPORT I_sp_teo(engine_t* engine);
-float TREL_EXPORT I_sp_esperado(engine_t* engine);
+double TREL_EXPORT I_sp_teo(engine_t* engine);
+double TREL_EXPORT I_sp_esperado(engine_t* engine);
 double TREL_EXPORT I_tot_teo(engine_t* engine);
-float TREL_EXPORT I_tot_esp(engine_t* engine);
+double TREL_EXPORT I_tot_esp(engine_t* engine);
 double TREL_EXPORT Delta_V_Esp(trel_rocket_t* rocket);
 
 #endif //RENDIMIENTOS_H
